requirements.txt
fastapi
uvicorn
GitPython
pydantic
python-multipart